detach service when underlying webview unmounts
