+++ src/containers/layout/AppLayoutContainer.js
@@ -42,6 +42,7 @@ export default @inject('stores', 'actions') @observer class AppLayoutContainer e
       setActive,
       handleIPCMessage,
       setWebviewReference,
+      detachService,
       openWindow,
       reorder,
       reload,
@@ -105,6 +106,7 @@ export default @inject('stores', 'actions') @observer class AppLayoutContainer e
         services={services.allDisplayedUnordered}
         handleIPCMessage={handleIPCMessage}
         setWebviewReference={setWebviewReference}
+        detachService={detachService}
         openWindow={openWindow}
         reload={reload}
         openSettings={openSettings}
@@ -160,6 +162,7 @@ AppLayoutContainer.wrappedComponent.propTypes = {
       toggleAudio: PropTypes.func.isRequired,
       handleIPCMessage: PropTypes.func.isRequired,
       setWebviewReference: PropTypes.func.isRequired,
+      detachService: PropTypes.func.isRequired,
       openWindow: PropTypes.func.isRequired,
       reloadUpdatedServices: PropTypes.func.isRequired,
       updateService: PropTypes.func.isRequired,
